+++ types/accountmapper_test.go
@@ -1,4 +1,4 @@
-package auth
+package types_test
 
 import (
 	"testing"
@@ -11,6 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/store"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	wire "github.com/cosmos/cosmos-sdk/wire"
+	"github.com/cosmos/cosmos-sdk/x/auth"
 )
 
 func setupMultiStore() (sdk.MultiStore, *sdk.KVStoreKey) {
@@ -25,11 +26,11 @@ func setupMultiStore() (sdk.MultiStore, *sdk.KVStoreKey) {
 func TestAccountMapperGetSet(t *testing.T) {
 	ms, capKey := setupMultiStore()
 	cdc := wire.NewCodec()
-	RegisterBaseAccount(cdc)
+	auth.RegisterBaseAccount(cdc)
 
 	// make context and mapper
 	ctx := sdk.NewContext(ms, abci.Header{}, false, nil)
-	mapper := NewAccountMapper(cdc, capKey, &BaseAccount{})
+	mapper := sdk.NewAccountMapper(cdc, capKey, &auth.BaseAccount{})
 
 	addr := sdk.Address([]byte("some-address"))
 
@@ -56,18 +57,4 @@ func TestAccountMapperGetSet(t *testing.T) {
 	acc = mapper.GetAccount(ctx, addr)
 	assert.NotNil(t, acc)
 	assert.Equal(t, newSequence, acc.GetSequence())
-}
-
-func TestAccountMapperSealed(t *testing.T) {
-	_, capKey := setupMultiStore()
-	cdc := wire.NewCodec()
-	RegisterBaseAccount(cdc)
-
-	// normal mapper exposes the wire codec
-	mapper := NewAccountMapper(cdc, capKey, &BaseAccount{})
-	assert.NotNil(t, mapper.WireCodec())
-
-	// seal mapper, should panic when we try to get the codec
-	mapperSealed := mapper.Seal()
-	assert.Panics(t, func() { mapperSealed.WireCodec() })
 }